--- pages-content/home/index.tsx
@@ -1,7 +1,0 @@
-import { FC } from 'react';
-
-export const HomePageContent: FC = () => {
-  return <div>
-    123
-  </div>;
-};
